refactor: update product price fields from Double to BigDecimal

+++ dukan/src/main/kotlin/net/thechance/dukan/api/dto/cart/CartResponse.kt
@@ -1,9 +1,10 @@
 package net.thechance.dukan.api.dto.cart
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigDecimal
 import java.util.*
 
 data class CartResponse(
     val id: UUID,
-    val totalPrice: Double,
+    val totalPrice: BigDecimal,
 )
